refactor: ProfilePage custom-hook, separation components
