+++ package.json
@@ -17,7 +17,6 @@
     "lodash": "^4.17.21",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-kakao-maps-sdk": "^1.1.27",
     "react-router-dom": "^7.0.1",
     "react-slick": "^0.30.2",
     "slick-carousel": "^1.8.1",
