+++ src/components/ui/home/ChefCardItem.jsx
@@ -9,7 +9,7 @@ const ChefCardItem = ({ info, color }) => {
           <img src={info.image_url} alt={`${info.name}`} className="w-full h-full object-cover scale-110 " />
         </div>
       </div>
-      <Link to={`/details/${info.id}`}>
+      <Link to={`/detail/${info.id}`}>
         <div className="absolute top-0 left-0 w-[320px] h-[320px] bg-[rgba(0,0,0,0.7)] rounded-[16px] px-10 text-[#fff] opacity-0 group-hover:opacity-100 transition-opacity duration-200 shadow-xl">
           <div className="w-full h-auto mt-20">
             <p className="text-[16px]">{info.description}</p>
